+++ osu.Game/OsuGameBase.cs
@@ -158,7 +158,7 @@ namespace osu.Game
                         Children = new Drawable[]
                         {
                             Cursor = new MenuCursor(),
-                            new OsuCursorContextMenu(Cursor) { Depth = -2 },
+                            new OsuContextMenuContainer(Cursor) { Depth = -2 },
                             new OsuTooltipContainer(Cursor) { Depth = -1 },
                         }
                     },
